moved ThemeButton component into Button folder
